feat(ordem-servico): adiciona comando criar ordem servico

+++ src/Application/Administrativo/Cliente/Commands/CreateCliente/CreateClienteCommandHandler.cs
@@ -28,7 +28,7 @@ public class CreateClienteCommandHandler(IClienteRepository clienteRepository)
         {
             Id = cliente.Id,
             Nome = cliente.Nome,
-            TipoDocumento = (int)cliente.TipoDocumento,
+            TipoDocumento = cliente.TipoDocumento,
             Documento = cliente.Documento
         };
     }
