+++ src/Application/Administrativo/Cliente/Commands/UpdateCliente/UpdateClienteCommandHandler.cs
@@ -31,7 +31,7 @@ public class UpdateClienteCommandHandler(IClienteRepository clienteRepository)
         {
             Id = cliente.Id,
             Nome = cliente.Nome,
-            TipoDocumento = (int)cliente.TipoDocumento,
+            TipoDocumento = cliente.TipoDocumento,
             Documento = cliente.Documento
         };
     }
